fixed bug in which users couldnt update their details
