+++ client/src/pages/MonthCardsLibrary/tests/index.test.jsx
@@ -1,5 +1,5 @@
 import CardsLibrary from '../CardsLibrary';
-import { render, screen } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import { expect, test } from 'vitest';
 import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
@@ -22,31 +22,32 @@ test('render cards and check buttons functionality', async () => {
 
   axios.get.mockResolvedValue({
     status: 200,
-    data: [
-      {
-        month: 8,
-        foodExpenses: 55,
-        subscriptionExpenses: 34,
-        transportExpenses: 13,
-        otherExpenses: 46,
-      },
-      {
-        month: 7,
-        foodExpenses: 54,
-        subscriptionExpenses: 30,
-        transportExpenses: 33,
-        otherExpenses: 22,
-      },
-      {
-        month: 6,
-      },
-      {
-        month: 3,
-      },
-      {
-        month: 5,
-      },
-    ],
+    data: {
+      cards: [
+        {
+          id: '1',
+          month: 8,
+          foodExpenses: 55,
+          subscriptionExpenses: 34,
+          transportExpenses: 13,
+          otherExpenses: 46,
+          currency: 'USD',
+        },
+        {
+          id: '2',
+          month: 7,
+          foodExpenses: 54,
+          subscriptionExpenses: 30,
+          transportExpenses: 33,
+          otherExpenses: 22,
+          currency: 'USD',
+        },
+        { id: '3', month: 6, currency: 'USD' },
+        { id: '4', month: 3, currency: 'USD' },
+        { id: '5', month: 5, currency: 'USD' },
+      ],
+      groupCards: {}, // if you don't care about summaryWidget for this test
+    },
   });
 
   render(
@@ -65,7 +66,6 @@ test('render cards and check buttons functionality', async () => {
 
   // Click next and check new state
   await user.click(nextButton);
-
   const cardPageTwo = await screen.findByTestId('5');
   expect(cardPageTwo).toBeInTheDocument();
   expect(screen.queryByTestId('8')).not.toBeInTheDocument();
@@ -76,4 +76,65 @@ test('render cards and check buttons functionality', async () => {
   // Re-query for the elements after going back
   expect(await screen.findByTestId('8')).toBeInTheDocument();
   expect(screen.queryByTestId('5')).not.toBeInTheDocument();
+});
+
+test('renders expenses text in card', async () => {
+  axios.get.mockResolvedValueOnce({
+    status: 200,
+    data: {
+      cards: [
+        {
+          id: '1',
+          month: 8,
+          foodExpenses: 55,
+          subscriptionExpenses: 34,
+          transportExpenses: 13,
+          otherExpenses: 46,
+          currency: 'USD',
+        },
+      ],
+      groupCards: {},
+    },
+  });
+
+  render(
+    <Provider store={store}>
+      <CardsLibrary />
+    </Provider>
+  );
+
+  const expenseText = await screen.findByText(/you have spent/i);
+  expect(expenseText).toHaveTextContent(
+    'You have spent 55 USD in groceries, 34 USD in subscriptions, 13 USD in transport, 46 USD and in misc!'
+  );
+});
+
+test('renders summary widget correctly', async () => {
+  axios.get.mockResolvedValueOnce({
+    status: 200,
+    data: {
+      cards: [],
+      groupCards: {
+        2025: [
+          { id: '10', month: 1 },
+          { id: '11', month: 2 },
+        ],
+      },
+    },
+  });
+
+  render(
+    <Provider store={store}>
+      <CardsLibrary />
+    </Provider>
+  );
+
+  await waitFor(() => {
+    expect(screen.queryByText(/loading/i)).not.toBeInTheDocument();
+  });
+  // Year should appear
+  expect(screen.getByText('2025')).toBeInTheDocument();
+
+  // Grouped months should appear
+  expect(screen.getAllByRole('link').length).toBeGreaterThan(0);
 });
