+++ client/src/pages/SignIn/tests/callAPI.test.jsx
@@ -31,10 +31,10 @@ test('successfull API call', async () => {
 
   const user = userEvent.setup();
 
-  await user.type(screen.getByLabelText(/Your Email/i), 'test@example.com');
+  await user.type(screen.getByLabelText(/Email/i), 'test@example.com');
   await user.type(screen.getByLabelText(/^Password$/i), 'password123');
 
-  const callAPIbutton = screen.getByRole('button', { name: /Submit/i });
+  const callAPIbutton = screen.getByRole('button', { name: /Sign in/i });
 
   await user.click(callAPIbutton);
 
@@ -61,10 +61,10 @@ test('failed API call err 400', async () => {
   );
   const user = userEvent.setup();
 
-  await user.type(screen.getByLabelText(/Your Email/i), 'test@example.com');
+  await user.type(screen.getByLabelText(/Email/i), 'test@example.com');
   await user.type(screen.getByLabelText(/^Password$/i), 'password123');
 
-  const callAPIbutton = screen.getByRole('button', { name: /Submit/i });
+  const callAPIbutton = screen.getByRole('button', { name: /Sign in/i });
 
   await user.click(callAPIbutton);
 
@@ -89,10 +89,10 @@ test('failed API call err 401', async () => {
   );
   const user = userEvent.setup();
 
-  await user.type(screen.getByLabelText(/Your Email/i), 'test@example.com');
+  await user.type(screen.getByLabelText(/Email/i), 'test@example.com');
   await user.type(screen.getByLabelText(/^Password$/i), 'password123');
 
-  const callAPIbutton = screen.getByRole('button', { name: /Submit/i });
+  const callAPIbutton = screen.getByRole('button', { name: /Sign in/i });
 
   await user.click(callAPIbutton);
