+++ client/src/pages/UserProfile/UserProfile.jsx
@@ -1,6 +1,6 @@
 import Axios from 'axios';
 import { useState } from 'react';
-import { useSelector } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
 import { Modal } from 'bootstrap';
 import { useEffect } from 'react';
 
@@ -14,7 +14,7 @@ import avatar from '../../assets/Anya.png';
 import AddDefaultItems from '../../components/AddDefaultItems/AddDefaultItems';
 import RecentCardWidget from '../../components/cardWidget/RecentCardWidget';
 import ExpensesSummary from '../ExpensesSummary/ExpensesSummary';
-import ExpenseInputFields from '../../components/addExpenseInline/ExpenseInputFields';
+import EditUserDetails from '../../components/EditUserDetails/EditUserDetails';
 
 const UserProfile = () => {
   const user = useSelector((state) => state.user); //
@@ -23,35 +23,18 @@ const UserProfile = () => {
   const currency = useSelector((state) => state.currency);
   const userAvatar = avatar || user?.avatar;
   const backendLink = import.meta.env.VITE_APP_API_URL;
+  const dispatch = useDispatch();
 
   const [profileEditMode, setProfileEditMode] = useState(false);
   const [defaultItemsEditMode, setDefaultItemsEditMode] = useState(false);
 
-  const [newName, setNewName] = useState(null);
-  const [newMail, setNewMail] = useState(null);
-
-  const [password, setPassword] = useState(null);
-  const [password2, setPassword2] = useState(null);
-
-  const [errMsgPassword, setErrMsgPassword] = useState(null);
-  const [isLoading, setIsLoading] = useState(false);
-  const [income, setIncome] = useState();
-  const [formData, setFormData] = useState({
-    newName: '',
-    newMail: '',
-    password: '',
-    password2: '',
-  });
-
-  const currentMonth = new Date().getUTCMonth();
   const currentYear = new Date().getFullYear();
 
-  const expensesYearIndex = user?.dataByYear.findIndex(
+  const expensesYearIndex = user?.dataByYear?.findIndex(
     (obj) => obj.year === currentYear
   );
 
-  const expensesYearSummary = user?.dataByYear[expensesYearIndex];
-
+  const expensesYearSummary = user?.dataByYear?.[expensesYearIndex] || {};
 
   const handleProfileEditingMode = () => {
     setProfileEditMode((prev) => !prev);
@@ -67,62 +50,6 @@ const UserProfile = () => {
     if (profileEditMode) {
       setProfileEditMode(false);
     }
-
-  };
-
-  const validatePassword = (pw1, pw2) => {
-    if (pw1 !== pw2) {
-      setErrMsgPassword(
-        <div className="errMsgPassword">
-          Please make sure both password fields are the same.
-        </div>
-      );
-      return false;
-    } else {
-      setErrMsgPassword(null);
-    }
-    return true;
-  };
-
-  const handleSave = async () => {
-    if (!validatePassword(password, password2)) {
-      return;
-    }
-    setProfileEditMode(false);
-
-    const modal = new Modal(document.getElementById('saveAlertModal'));
-    modal.show();
-
-    try {
-      if (profileEditMode) {
-        const response = await Axios.patch(
-          `${backendLink}/users/${userid}/update`,
-          {
-            name: formData.newName || undefined,
-            email: formData.newMail || undefined,
-            password: formData.password || undefined,
-          },
-          { headers: { Authorization: `Bearer ${token}` } }
-        );
-
-        if (response.status === 200) {
-          const { name, email } = response.data;
-          dispatch(setLogin({ name, email }));
-          setProfileEditMode(false);
-          setFormData({
-            newName: '',
-            newMail: '',
-            password: '',
-            password2: '',
-          });
-
-          modal.show();
-        }
-      }
-
-    } catch (error) {
-      console.log(error);
-    }
   };
 
   const renderProfileInfo = () => {
@@ -130,17 +57,19 @@ const UserProfile = () => {
       const allitems = user.defaultItems;
       const tableData = [];
 
-      Object.entries(allitems).forEach(([categoryName, categoryData]) => {
-        if (categoryData.items && categoryData.items.length > 0) {
-          categoryData.items.forEach((item, index) => {
-            tableData.push({
-              category: index === 0 ? categoryData.name : '', // Only show category on first item
-              description: item?.description || 'No description',
-              price: item?.price || 0,
+      Object.entries(allitems || {})?.forEach(
+        ([categoryName, categoryData]) => {
+          if (categoryData.items && categoryData.items.length > 0) {
+            categoryData.items.forEach((item, index) => {
+              tableData.push({
+                category: index === 0 ? categoryData.name : '', // Only show category on first item
+                description: item?.description || 'No description',
+                price: item?.price || 0,
+              });
             });
-          });
+          }
         }
-      });
+      );
 
       return (
         <div className="profile-info-div">
@@ -188,99 +117,12 @@ const UserProfile = () => {
     }
   };
 
-  // Render profile edit form
-  const renderProfileEditForm = () => (
-    <div className="form-grid editing-profile-div">
-      <div className="d-flex flex-row align-items-center ">
-        <div className="form-floating flex-grow-1">
-          <i className="fas fa-envelope fa-lg me-3 fa-fw"></i>
-          <input
-            onChange={(e) => setNewName(e.target.value)}
-            type="text"
-            id="form-name"
-            className="form-control"
-          />
-          <label className="form-label" htmlFor="form-name">
-            Your Name
-          </label>
-        </div>
-        <i
-          className="info-warning"
-          data-bs-toggle="tooltip"
-          data-bs-placement="top"
-          title="your name can only be updated once every 3 months"
-        >
-          <FontAwesomeIcon icon={faCircleInfo} />
-        </i>
-      </div>
-
-      <div className="d-flex flex-row align-items-center ">
-        <div className="form-floating flex-fill mb-0">
-          <i className="fas fa-envelope fa-lg me-3 fa-fw"></i>
-          <input
-            onChange={(e) => setNewMail(e.target.value)}
-            type="email"
-            id="form-email"
-            className="form-control"
-          />
-          <label className="form-label" htmlFor="form-email">
-            Your Email
-          </label>
-        </div>
-        <i
-          className="info-warning"
-          data-bs-toggle="tooltip"
-          data-bs-placement="top"
-          title="your mail can only be updated once every 6 months"
-        >
-          <FontAwesomeIcon icon={faCircleInfo} />
-        </i>
-      </div>
-
-      <div className="d-flex flex-row align-items-center ">
-        <div className="form-floating flex-fill mb-0">
-          <i className="fas fa-envelope fa-lg me-3 fa-fw"></i>
-          <input
-            onChange={(e) => setPassword(e.target.value)}
-            type="password"
-            id="form-password"
-            className="form-control"
-          />
-          <label className="form-label" htmlFor="form-password">
-            New password
-          </label>
-        </div>
-      </div>
-
-      <div className="d-flex flex-row align-items-center mb-4">
-        <div className="form-floating flex-fill mb-0">
-          <i className="fas fa-key fa-lg me-3 fa-fw"></i>
-          <input
-            onChange={(e) => setPassword2(e.target.value)}
-            type="password"
-            id="form-repeat-pw"
-            className="form-control"
-          />
-          <label className="form-label" htmlFor="form-repeat-pw">
-            Repeat your new password
-          </label>
-        </div>
-      </div>
-
-      {errMsgPassword}
-
-      <button className="btn btn-dark" onClick={handleSave}>
-        Save
-      </button>
-    </div>
-  );
-
   const renderDynamicContent = () => {
     if (defaultItemsEditMode) {
       return <AddDefaultItems />;
     }
     if (profileEditMode) {
-      return renderProfileEditForm();
+      return <EditUserDetails />;
     }
     return renderProfileInfo();
   };
