+++ client/src/pages/UserProfile/tests/callAPI.test.jsx
@@ -69,14 +69,12 @@ test('input values reach the backend and update it', async () => {
   await user.click(saveButton);
 
   expect(axios.patch).toHaveBeenCalledWith(
-    `http://localhost:3000/users/${mockUserId}/update`,
+    `http://localhost:3000/users/update/${mockUserId}`,
     {
-      userId: mockUserId,
-      token: mockToken,
       name: 'Alicia',
       email: 'mail@mail',
       password: 'p12345',
-    }
+    }, { headers: { Authorization: `Bearer ${mockToken}` } }
   );
 
   expect(
